moving layer into constructor
