+++ dynamodb/src/test/scala/zio/dynamodb/ExecutorSpec.scala
@@ -1,9 +1,23 @@
 package zio.dynamodb
 
-import zio.Chunk
+import io.github.vigoo.zioaws.dynamodb.DynamoDb
+import io.github.vigoo.zioaws.dynamodb.DynamoDb.DynamoDbMock
+import io.github.vigoo.zioaws.dynamodb.model.{
+  AttributeValue,
+  BatchGetItemResponse,
+  BatchWriteItemResponse,
+  KeysAndAttributes
+}
+import zio.clock.Clock
+import zio.{ Chunk, ULayer, ZLayer }
 import zio.dynamodb.DynamoDBQuery._
+
+import scala.collection.immutable.{ Map => ScalaMap }
+import zio.duration._
 import zio.test.Assertion._
-import zio.test.{ assert, DefaultRunnableSpec, TestAspect, ZSpec }
+import zio.test.environment.{ testEnvironment, Live, TestClock }
+import zio.test.mock.Expectation.value
+import zio.test.{ assert, Annotations, DefaultRunnableSpec, TestAspect, ZSpec }
 
 object ExecutorSpec extends DefaultRunnableSpec with DynamoDBFixtures {
 
@@ -18,7 +32,10 @@ object ExecutorSpec extends DefaultRunnableSpec with DynamoDBFixtures {
   )
 
   override def spec: ZSpec[Environment, Failure] =
-    suite("Batching")(crudSuite, scanAndQuerySuite, batchingSuite).provideLayer(DynamoDBExecutor.test)
+    suite("Executor spec")(
+      suite("Batching")(crudSuite, scanAndQuerySuite, batchingSuite).provideLayer(DynamoDBExecutor.test),
+      batchRetries
+    )
 
   private val crudSuite = suite("single Item CRUD suite")(
     testM("getItem") {
@@ -160,4 +177,152 @@ object ExecutorSpec extends DefaultRunnableSpec with DynamoDBFixtures {
     } @@ beforeAddTable1AndTable2
   )
 
+  private val mockBatches     = "mockBatches"
+  private val itemOne         = Item("k1" -> "v1")
+  private val firstGetRequest =
+    DynamoDBExecutorImpl.generateBatchGetItemRequest(
+      BatchGetItem(
+        ScalaMap(
+          TableName(mockBatches) -> BatchGetItem.TableGet(
+            keysSet = Set(PrimaryKey("k1" -> "v1"), PrimaryKey("k1" -> "v2")),
+            projectionExpressionSet = Set.empty
+          )
+        )
+      )
+    )
+
+  private val retryGetRequest =
+    DynamoDBExecutorImpl.generateBatchGetItemRequest(
+      BatchGetItem(
+        ScalaMap(
+          TableName(mockBatches) -> BatchGetItem.TableGet(
+            keysSet = Set(PrimaryKey("k1" -> "v2")),
+            projectionExpressionSet = Set.empty
+          )
+        )
+      )
+    )
+
+  private val batchWriteRequest = BatchWriteItem(
+    MapOfSet.empty[TableName, BatchWriteItem.Write] + (
+      (
+        TableName(mockBatches),
+        BatchWriteItem.Put(itemOne)
+      )
+    ),
+    exponentialBackoff = 0.seconds,
+    retryAttempts = 1
+  )
+
+  private val firstWriteRequest =
+    DynamoDBExecutorImpl.generateBatchWriteItem(
+      batchWriteRequest
+    )
+
+  private val mockedBatchGet: ULayer[DynamoDb] = DynamoDbMock
+    .BatchGetItem(
+      equalTo(firstGetRequest),
+      value(
+        BatchGetItemResponse(
+          responses = Some(
+            ScalaMap(
+              mockBatches -> List(
+                ScalaMap("k1" -> AttributeValue(s = Some("v1")))
+              )
+            )
+          ),
+          unprocessedKeys = Some(
+            ScalaMap(
+              mockBatches -> KeysAndAttributes(
+                keys = List(ScalaMap("k1" -> AttributeValue(s = Some("v2"))))
+              )
+            )
+          )
+        ).asReadOnly
+      )
+    ) ++ DynamoDbMock
+    .BatchGetItem(
+      equalTo(retryGetRequest),
+      value(
+        BatchGetItemResponse(
+          responses = Some(
+            ScalaMap(
+              mockBatches -> List(
+                ScalaMap(
+                  "k1" -> AttributeValue(s = Some("v2")),
+                  "k2" -> AttributeValue(s = Some("v23"))
+                )
+              )
+            )
+          ),
+          unprocessedKeys = None
+        ).asReadOnly
+      )
+    )
+
+  private val batchGetSuite =
+    suite("retry batch gets")(
+      testM("should retry when there are unprocessed keys") {
+        for {
+          response <- BatchGetItem(
+                        requestItems = ScalaMap(
+                          TableName(mockBatches) -> BatchGetItem.TableGet(
+                            Set(PrimaryKey("k1" -> "v1"), PrimaryKey("k1" -> "v2")),
+                            Set.empty
+                          )
+                        ),
+                        exponentialBackoff = 0.seconds
+                      ).execute
+        } yield assert(response.responses.get(TableName(mockBatches)))(
+          equalTo(Some(Set(itemOne, Item("k1" -> "v2", "k2" -> "v23"))))
+        )
+      }
+    )
+
+  private val mockedBatchWrite: ULayer[DynamoDb] = DynamoDbMock
+    .BatchWriteItem(
+      equalTo(firstWriteRequest),
+      value(
+        BatchWriteItemResponse(
+          unprocessedItems = Some(
+            ScalaMap(
+              mockBatches -> Set(
+                DynamoDBExecutorImpl.batchItemWriteToZIOAwsWriteRequest(BatchWriteItem.Put(itemOne))
+              )
+            )
+          )
+        ).asReadOnly
+      )
+    )
+    .atMost(2)
+
+  private val batchWriteSuite =
+    suite("retry batch writes")(
+      testM("should retry when there are unprocessed items") {
+        for {
+          response <- batchWriteRequest.execute
+
+        } yield assert(response.unprocessedItems)(
+          equalTo(
+            Some(
+              MapOfSet
+                .empty[TableName, BatchWriteItem.Write] + (
+                (
+                  TableName(mockBatches),
+                  BatchWriteItem.Put(itemOne)
+                )
+              )
+            )
+          )
+        )
+      }
+    )
+
+  private val clockLayer: ZLayer[Any, Nothing, Clock with TestClock] =
+    testEnvironment >>> ((Annotations.live ++ Live.default) >>> TestClock.default)
+
+  private val batchRetries = suite("Batch retries")(
+    batchGetSuite.provideLayer((mockedBatchGet ++ clockLayer) >>> DynamoDBExecutor.live),
+    batchWriteSuite.provideLayer((mockedBatchWrite ++ clockLayer) >>> DynamoDBExecutor.live)
+  )
 }
